version con mejora
